Turned DataGraph into a user control Added Axis numbers (partial)
AttribMod editor: reindexing guide numbers, starting from 1
