+++ midsControls/midsControls.csproj
@@ -96,6 +96,12 @@
     <Compile Include="EditableLabel.Designer.cs">
       <DependentUpon>EditableLabel.cs</DependentUpon>
     </Compile>
+    <Compile Include="ctlDataGraph.cs">
+      <SubType>Component</SubType>
+    </Compile>
+    <Compile Include="ctlDataGraph.Designer.cs">
+      <DependentUpon>ctlDataGraph.cs</DependentUpon>
+    </Compile>
   </ItemGroup>
   <ItemGroup>
     <EmbeddedResource Include="ctlDamageDisplay.resources" />
